+++ familyhelper-clannad-node/src/main/java/com/dwarfeng/familyhelper/clannad/node/configuration/CacheConfiguration.java
@@ -1,9 +1,7 @@
 package com.dwarfeng.familyhelper.clannad.node.configuration;
 
-import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonGenderTypeIndicator;
 import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonProfile;
 import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonUser;
-import com.dwarfeng.familyhelper.clannad.stack.bean.entity.GenderTypeIndicator;
 import com.dwarfeng.familyhelper.clannad.stack.bean.entity.Profile;
 import com.dwarfeng.familyhelper.clannad.stack.bean.entity.User;
 import com.dwarfeng.subgrade.impl.bean.DozerBeanTransformer;
@@ -24,8 +22,6 @@ public class CacheConfiguration {
 
     @Value("${cache.prefix.entity.profile}")
     private String profilePrefix;
-    @Value("${cache.prefix.entity.gender_type_indicator}")
-    private String genderTypeIndicatorPrefix;
     @Value("${cache.prefix.entity.user}")
     private String userPrefix;
 
@@ -41,17 +37,6 @@ public class CacheConfiguration {
                 (RedisTemplate<String, FastJsonProfile>) template,
                 new StringIdStringKeyFormatter(profilePrefix),
                 new DozerBeanTransformer<>(Profile.class, FastJsonProfile.class, mapper)
-        );
-    }
-
-    @Bean
-    @SuppressWarnings("unchecked")
-    public RedisBatchBaseCache<StringIdKey, GenderTypeIndicator, FastJsonGenderTypeIndicator>
-    genderTypeIndicatorRedisBatchBaseCache() {
-        return new RedisBatchBaseCache<>(
-                (RedisTemplate<String, FastJsonGenderTypeIndicator>) template,
-                new StringIdStringKeyFormatter(genderTypeIndicatorPrefix),
-                new DozerBeanTransformer<>(GenderTypeIndicator.class, FastJsonGenderTypeIndicator.class, mapper)
         );
     }