+++ familyhelper-clannad-node/src/main/java/com/dwarfeng/familyhelper/clannad/node/configuration/FastJsonConfiguration.java
@@ -1,7 +1,6 @@
 package com.dwarfeng.familyhelper.clannad.node.configuration;
 
 import com.alibaba.fastjson.parser.ParserConfig;
-import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonGenderTypeIndicator;
 import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonProfile;
 import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonUser;
 import org.slf4j.Logger;
@@ -17,7 +16,6 @@ public class FastJsonConfiguration {
         LOGGER.info("正在配置 FastJson autotype 白名单");
         //实体对象。
         ParserConfig.getGlobalInstance().addAccept(FastJsonProfile.class.getCanonicalName());
-        ParserConfig.getGlobalInstance().addAccept(FastJsonGenderTypeIndicator.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonUser.class.getCanonicalName());
         LOGGER.debug("FastJson autotype 白名单配置完毕");
     }